fraud_detection_system/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==21.6b0
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
DateTime==4.3
Django==3.1.7
django-cors-headers==3.7.0
django-crontab==0.7.1
django-filter==2.4.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
idna==2.10
importlib==1.0.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonfield==3.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
PyJWT==1.7.1
pyparsing==2.4.7
python-decouple==3.4
pytz==2021.1
regex==2021.4.4
requests==2.25.1
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.1
swapper==1.1.2.post1
toml==0.10.2
twilio==6.62.0
uritemplate==3.0.1
urllib3==1.26.4
zope.interface==5.4.0
